refactor: Extract info item data structures into individual actionable structs.
