+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/faq/FetchFAQUseCase.kt
@@ -1,7 +1,7 @@
 package no.usn.mob3000.domain.usecase.content.faq
 
 import no.usn.mob3000.data.repository.content.FAQRepository
-import no.usn.mob3000.domain.model.FAQData
+import no.usn.mob3000.domain.model.content.FAQData
 
 /**
  * Use case for fetching FAQ.
