+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/news/FetchNewsUseCase.kt
@@ -1,7 +1,7 @@
 package no.usn.mob3000.domain.usecase.content.news
 
 import no.usn.mob3000.data.repository.content.NewsRepository
-import no.usn.mob3000.domain.model.NewsData
+import no.usn.mob3000.domain.model.content.NewsData
 
 /**
  * Use case for fetching news.
