+++ app/src/main/java/no/usn/mob3000/ui/screens/info/docs/UpdateDocumentationScreen.kt
@@ -1,8 +1,7 @@
 package no.usn.mob3000.ui.screens.info.docs
 
 import androidx.compose.runtime.*
-import no.usn.mob3000.domain.model.DocsData
-import no.usn.mob3000.domain.model.FAQData
+import no.usn.mob3000.domain.model.content.DocsData
 import no.usn.mob3000.ui.components.info.ContentEditor
 
 /**
