+++ app/src/main/java/no/usn/mob3000/ui/screens/info/faq/FAQScreen.kt
@@ -1,15 +1,15 @@
 package no.usn.mob3000.ui.screens.info.faq
 
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import no.usn.mob3000.Viewport
 import no.usn.mob3000.ui.theme.DefaultButton
-import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.*
 import kotlinx.coroutines.flow.StateFlow
-import no.usn.mob3000.domain.model.FAQData
+import no.usn.mob3000.domain.model.content.FAQData
 import no.usn.mob3000.ui.components.info.ContentItem
 import no.usn.mob3000.ui.components.info.PaddedLazyColumn
