+++ app/src/main/java/no/usn/mob3000/ui/screens/info/faq/UpdateFAQScreen.kt
@@ -1,8 +1,7 @@
 package no.usn.mob3000.ui.screens.info.faq
 
 import androidx.compose.runtime.*
-import no.usn.mob3000.Destination
-import no.usn.mob3000.domain.model.FAQData
+import no.usn.mob3000.domain.model.content.FAQData
 import no.usn.mob3000.ui.components.info.ContentEditor
 
 /**
